move useLegacyMixinAp inside MixinExtension
