+++ src/main/java/net/fabricmc/loom/extension/MixinExtensionApiImpl.java
@@ -24,6 +24,8 @@
 
 package net.fabricmc.loom.extension;
 
+import java.util.Objects;
+
 import org.gradle.api.Action;
 import org.gradle.api.InvalidUserDataException;
 import org.gradle.api.Project;
@@ -36,13 +38,25 @@ import org.gradle.api.tasks.util.PatternSet;
 import net.fabricmc.loom.api.MixinExtensionAPI;
 
 public abstract class MixinExtensionApiImpl implements MixinExtensionAPI {
-	protected abstract Project getProject();
+	protected final Project project;
+	protected final Property<Boolean> useMixinAp;
+
+	public MixinExtensionApiImpl(Project project) {
+		this.project = Objects.requireNonNull(project);
+		this.useMixinAp = project.getObjects().property(Boolean.class)
+				.convention(false);
+	}
 
 	protected final PatternSet add0(SourceSet sourceSet, String refmapName) {
-		return add0(sourceSet, getProject().provider(() -> refmapName));
+		return add0(sourceSet, project.provider(() -> refmapName));
 	}
 
 	protected abstract PatternSet add0(SourceSet sourceSet, Provider<String> refmapName);
+
+	@Override
+	public Property<Boolean> getUseLegacyMixinAp() {
+		return useMixinAp;
+	}
 
 	@Override
 	public void add(SourceSet sourceSet, String refmapName, Action<PatternSet> action) {
@@ -57,7 +71,7 @@ public abstract class MixinExtensionApiImpl implements MixinExtensionAPI {
 
 	@Override
 	public void add(String sourceSetName, String refmapName, Action<PatternSet> action) {
-		add(sourceSetName, getProject().provider(() -> refmapName), action);
+		add(sourceSetName, project.provider(() -> refmapName), action);
 	}
 
 	public void add(String sourceSetName, Provider<String> refmapName, Action<PatternSet> action) {
@@ -96,7 +110,7 @@ public abstract class MixinExtensionApiImpl implements MixinExtensionAPI {
 
 	private SourceSet resolveSourceSet(String sourceSetName) {
 		// try to find sourceSet with name sourceSetName in this project
-		SourceSet sourceSet = getProject().getConvention().getPlugin(JavaPluginConvention.class)
+		SourceSet sourceSet = project.getConvention().getPlugin(JavaPluginConvention.class)
 				.getSourceSets().findByName(sourceSetName);
 
 		if (sourceSet == null) {
@@ -109,13 +123,8 @@ public abstract class MixinExtensionApiImpl implements MixinExtensionAPI {
 	// This is here to ensure that LoomGradleExtensionApiImpl compiles without any unimplemented methods
 	private final class EnsureCompile extends MixinExtensionApiImpl {
 		private EnsureCompile() {
-			super();
+			super(null);
 			throw new RuntimeException();
-		}
-
-		@Override
-		protected Project getProject() {
-			throw new RuntimeException("Yeah... something is really wrong");
 		}
 
 		@Override
